refactor remove utils
