+++ examples/solana/oyster/actions/account.ts
@@ -10,7 +10,7 @@ import {
   SPL_ASSOCIATED_TOKEN_ACCOUNT_PROGRAM_ID,
   TOKEN_PROGRAM_ID,
   WRAPPED_SOL_MINT,
-} from '../utils/ids';
+} from '../../utils/ids';
 // import { programIds } from '../utils/programIds';
 import { TokenAccount } from '../models/account';
 // import { cache } from '../contexts/accounts/cache';
