+++ examples/solana/oyster/actions/vault.ts
@@ -6,7 +6,7 @@ import {
 // import { programIds } from '../utils/programIds';
 import { deserializeUnchecked, serialize } from 'borsh';
 import BN from 'bn.js';
-import { findProgramAddress, StringPublicKey, toPublicKey } from '../utils';
+import { findProgramAddress, StringPublicKey, toPublicKey } from '../../utils';
 import { programIds } from '../../utils';
 
 export const VAULT_PREFIX = 'vault';
