+++ examples/solana/oyster/contexts/accounts/accounts.ts
@@ -1,7 +1,7 @@
 import { AccountInfo, Connection, PublicKey } from '@solana/web3.js';
 import { AccountLayout, MintInfo, u64 } from '@solana/spl-token';
 import { TokenAccount } from '../../models';
-import { StringPublicKey, WRAPPED_SOL_MINT } from '../../utils/ids';
+import { StringPublicKey, WRAPPED_SOL_MINT } from '../../../utils/ids';
 // import { programIds } from '../../utils/programIds';
 // import { genericCache, cache } from './cache';
 import { deserializeAccount } from './deserialize';
