+++ examples/solana/oyster/contexts/meta/index.ts
@@ -1,4 +1,3 @@
-export * from './meta';
 export * from './isMetadataPartOfStore';
 export * from './loadAccounts';
 export * from './onChangeAccount';
