+++ examples/solana/oyster/contexts/meta/loadAccounts.ts
@@ -5,7 +5,7 @@ import {
   StringPublicKey,
   toPublicKey,
   VAULT_ID,
-} from '../../utils/ids';
+} from '../../../utils/ids';
 import { MAX_WHITELISTED_CREATOR_SIZE } from '../../models';
 import {
   getEdition,
@@ -38,7 +38,7 @@ import { ParsedAccount } from '../accounts/types';
 import { getEmptyMetaState } from './getEmptyMetaState';
 import { getMultipleAccounts } from '../accounts/getMultipleAccounts';
 import { getProgramAccounts } from './web3';
-import { createPipelineExecutor } from '../../utils/createPipelineExecutor';
+import { createPipelineExecutor } from '../../../utils/createPipelineExecutor';
 
 export const USE_SPEED_RUN = false;
 const WHITELISTED_METADATA = ['98vYFjBYS9TguUMWQRPjy2SZuxKuUMcqR4vnQiLjZbte'];
