--- examples/solana/oyster/contexts/meta/meta.ts
@@ -1,10 +1,0 @@
-import { queryExtendedMetadata } from './queryExtendedMetadata';
-import { subscribeAccountsChange } from './subscribeAccountsChange';
-import { getEmptyMetaState } from './getEmptyMetaState';
-import {
-  limitedLoadAccounts,
-  loadAccounts,
-  USE_SPEED_RUN,
-} from './loadAccounts';
-import { MetaContextState, MetaState } from './types';
-import { AuctionData, BidderMetadata, BidderPot } from '../../actions';
