--- examples/solana/oyster/contexts/wallet.css
@@ -1,6 +1,0 @@
-p {
-  color: #6d6d6d;
-  font-family: Inter, sans-serif;
-  font-style: normal;
-  font-size: 0.9rem;
-}
