+++ examples/solana/oyster/contexts/wallet.ts
@@ -1,19 +1,4 @@
-import { WalletAdapter, WalletError } from '@solana/wallet-adapter-base';
-import {
-  useWallet,
-  WalletProvider as BaseWalletProvider,
-} from '@solana/wallet-adapter-react';
-import {
-  getLedgerWallet,
-  getMathWallet,
-  getPhantomWallet,
-  getSolflareWallet,
-  getSolletWallet,
-  getSolongWallet,
-  getTorusWallet,
-  WalletName,
-} from '@solana/wallet-adapter-wallets';
-import { Button } from 'antd';
+import { WalletAdapter } from '@solana/wallet-adapter-base';
 
 export interface WalletModalContextState {
   visible: boolean;
