+++ examples/solana/oyster/models/metaplex/decommissionAuctionManager.ts
@@ -2,7 +2,7 @@ import { SYSVAR_CLOCK_PUBKEY, TransactionInstruction } from '@solana/web3.js';
 import { serialize } from 'borsh';
 
 import { DecommissionAuctionManagerArgs, SCHEMA } from '.';
-import { programIds, StringPublicKey, toPublicKey } from '../../utils';
+import { programIds, StringPublicKey, toPublicKey } from '../../../utils';
 
 export async function decommissionAuctionManager(
   auctionManager: StringPublicKey,
