+++ examples/solana/oyster/models/metaplex/deprecatedValidateSafetyDepositBoxV1.ts
@@ -7,7 +7,7 @@ import { serialize } from 'borsh';
 
 import { getAuctionKeys, getOriginalAuthority, SCHEMA } from '.';
 import { programIds } from '../../../utils';
-import { StringPublicKey, toPublicKey } from '../../utils';
+import { StringPublicKey, toPublicKey } from '../../../utils';
 
 import {
   getSafetyDepositBoxValidationTicket,
