+++ examples/solana/oyster/models/metaplex/initAuctionManagerV2.ts
@@ -14,7 +14,7 @@ import {
   TupleNumericType,
 } from '.';
 import { programIds } from '../../../utils';
-import { StringPublicKey, toPublicKey } from '../../utils';
+import { StringPublicKey, toPublicKey } from '../../../utils';
 
 export async function initAuctionManagerV2(
   vault: StringPublicKey,
