+++ examples/solana/oyster/models/metaplex/setStore.ts
@@ -2,7 +2,8 @@ import { SYSVAR_RENT_PUBKEY, TransactionInstruction } from '@solana/web3.js';
 import { serialize } from 'borsh';
 
 import { SCHEMA, SetStoreArgs } from '.';
-import { programIds, StringPublicKey, toPublicKey } from '../../utils';
+import { programIds } from '../../../utils';
+import { StringPublicKey, toPublicKey } from '../../../utils';
 
 export async function setStore(
   isPublic: boolean,
