--- examples/solana/oyster/utils/isValidHttpUrl.ts
@@ -1,7 +1,0 @@
-export function isValidHttpUrl(text: string) {
-  if (text.startsWith('http:') || text.startsWith('https:')) {
-    return true;
-  }
-
-  return false;
-}
